backend-plugin-api: document core service refs
