+++ .changeset/rich-clouds-smile.md
@@ -1,0 +1,5 @@
+---
+'@backstage/backend-plugin-api': patch
+---
+
+Documented `coreServices` an all of its members.
