+++ packages/backend-plugin-api/src/services/definitions/PermissionsService.ts
@@ -14,18 +14,10 @@
  * limitations under the License.
  */
 
-import { Config } from '@backstage/config';
-import { createServiceRef } from '../system/types';
+import {
+  PermissionAuthorizer,
+  PermissionEvaluator,
+} from '@backstage/plugin-permission-common';
 
-/**
- * @public
- */
-export type ConfigService = Config;
-
-/**
- * @public
- */
-export const configServiceRef = createServiceRef<ConfigService>({
-  id: 'core.root.config',
-  scope: 'root',
-});
+/** @public */
+export type PermissionsService = PermissionEvaluator | PermissionAuthorizer;
