+++ packages/backend-plugin-api/src/services/definitions/RootLifecycleService.ts
@@ -14,32 +14,7 @@
  * limitations under the License.
  */
 
-import { createServiceRef } from '../system/types';
+import { LifecycleService } from './LifecycleService';
 
-/**
- * @public
- **/
-export type LifecycleServiceShutdownHook = {
-  fn: () => void | Promise<void>;
-
-  /** Labels to help identify the shutdown hook */
-  labels?: Record<string, string>;
-};
-
-/**
- * @public
- **/
-export interface LifecycleService {
-  /**
-   * Register a function to be called when the backend is shutting down.
-   */
-  addShutdownHook(options: LifecycleServiceShutdownHook): void;
-}
-
-/**
- * @public
- */
-export const lifecycleServiceRef = createServiceRef<LifecycleService>({
-  id: 'core.lifecycle',
-  scope: 'plugin',
-});
+/** @public */
+export type RootLifecycleService = LifecycleService;
