+++ packages/backend-plugin-api/src/services/definitions/RootLoggerService.ts
@@ -14,15 +14,7 @@
  * limitations under the License.
  */
 
-import { createServiceRef } from '../system/types';
-import { PluginCacheManager } from '@backstage/backend-common';
+import { LoggerService } from './LoggerService';
 
 /** @public */
-export type CacheService = PluginCacheManager;
-
-/**
- * @public
- */
-export const cacheServiceRef = createServiceRef<CacheService>({
-  id: 'core.cache',
-});
+export type RootLoggerService = LoggerService;
