+++ packages/backend-plugin-api/src/services/definitions/SchedulerService.ts
@@ -14,19 +14,7 @@
  * limitations under the License.
  */
 
-import { createServiceRef } from '../system/types';
-import { Handler } from 'express';
+import { PluginTaskScheduler } from '@backstage/backend-tasks';
 
-/**
- * @public
- */
-export interface HttpRouterService {
-  use(handler: Handler): void;
-}
-
-/**
- * @public
- */
-export const httpRouterServiceRef = createServiceRef<HttpRouterService>({
-  id: 'core.httpRouter',
-});
+/** @public */
+export type SchedulerService = PluginTaskScheduler;
