+++ packages/backend-plugin-api/src/services/definitions/coreServices.ts
@@ -14,17 +14,165 @@
  * limitations under the License.
  */
 
-export { configServiceRef as config } from './configServiceRef';
-export { httpRouterServiceRef as httpRouter } from './httpRouterServiceRef';
-export { loggerServiceRef as logger } from './loggerServiceRef';
-export { urlReaderServiceRef as urlReader } from './urlReaderServiceRef';
-export { cacheServiceRef as cache } from './cacheServiceRef';
-export { databaseServiceRef as database } from './databaseServiceRef';
-export { discoveryServiceRef as discovery } from './discoveryServiceRef';
-export { tokenManagerServiceRef as tokenManager } from './tokenManagerServiceRef';
-export { permissionsServiceRef as permissions } from './permissionsServiceRef';
-export { schedulerServiceRef as scheduler } from './schedulerServiceRef';
-export { rootLifecycleServiceRef as rootLifecycle } from './rootLifecycleServiceRef';
-export { rootLoggerServiceRef as rootLogger } from './rootLoggerServiceRef';
-export { pluginMetadataServiceRef as pluginMetadata } from './pluginMetadataServiceRef';
-export { lifecycleServiceRef as lifecycle } from './lifecycleServiceRef';
+import { createServiceRef } from '../system';
+
+/**
+ * All core services references
+ *
+ * @public
+ */
+export namespace coreServices {
+  /**
+   * The service reference for the plugin scoped {@link CacheService}.
+   *
+   * @public
+   */
+  export const cache = createServiceRef<import('./CacheService').CacheService>({
+    id: 'core.cache',
+  });
+
+  /**
+   * The service reference for the root scoped {@link ConfigService}.
+   *
+   * @public
+   */
+  export const config = createServiceRef<
+    import('./ConfigService').ConfigService
+  >({
+    id: 'core.root.config',
+    scope: 'root',
+  });
+
+  /**
+   * The service reference for the plugin scoped {@link DatabaseService}.
+   *
+   * @public
+   */
+  export const database = createServiceRef<
+    import('./DatabaseService').DatabaseService
+  >({
+    id: 'core.database',
+  });
+
+  /**
+   * The service reference for the plugin scoped {@link DiscoveryService}.
+   *
+   * @public
+   */
+  export const discovery = createServiceRef<
+    import('./DiscoveryService').DiscoveryService
+  >({
+    id: 'core.discovery',
+  });
+
+  /**
+   * The service reference for the plugin scoped {@link HttpRouterService}.
+   *
+   * @public
+   */
+  export const httpRouter = createServiceRef<
+    import('./HttpRouterService').HttpRouterService
+  >({ id: 'core.httpRouter' });
+
+  /**
+   * The service reference for the plugin scoped {@link LifecycleService}.
+   *
+   * @public
+   */
+  export const lifecycle = createServiceRef<
+    import('./LifecycleService').LifecycleService
+  >({
+    id: 'core.lifecycle',
+    scope: 'plugin',
+  });
+
+  /**
+   * The service reference for the plugin scoped {@link LoggerService}.
+   *
+   * @public
+   */
+  export const logger = createServiceRef<
+    import('./LoggerService').LoggerService
+  >({
+    id: 'core.logger',
+  });
+
+  /**
+   * The service reference for the plugin scoped {@link PermissionsService}.
+   *
+   * @public
+   */
+  export const permissions = createServiceRef<
+    import('./PermissionsService').PermissionsService
+  >({
+    id: 'core.permissions',
+  });
+
+  /**
+   * The service reference for the plugin scoped {@link PluginMetadataService}.
+   *
+   * @public
+   */
+  export const pluginMetadata = createServiceRef<
+    import('./PluginMetadataService').PluginMetadataService
+  >({
+    id: 'core.plugin-metadata',
+  });
+
+  /**
+   * The service reference for the root scoped {@link RootLifecycleService}.
+   *
+   * @public
+   */
+  export const rootLifecycle = createServiceRef<
+    import('./RootLifecycleService').RootLifecycleService
+  >({
+    id: 'core.rootLifecycle',
+    scope: 'root',
+  });
+
+  /**
+   * The service reference for the root scoped {@link RootLoggerService}.
+   *
+   * @public
+   */
+  export const rootLogger = createServiceRef<
+    import('./RootLoggerService').RootLoggerService
+  >({
+    id: 'core.root.logger',
+    scope: 'root',
+  });
+
+  /**
+   * The service reference for the plugin scoped {@link SchedulerService}.
+   *
+   * @public
+   */
+  export const scheduler = createServiceRef<
+    import('./SchedulerService').SchedulerService
+  >({
+    id: 'core.scheduler',
+  });
+
+  /**
+   * The service reference for the plugin scoped {@link TokenManagerService}.
+   *
+   * @public
+   */
+  export const tokenManager = createServiceRef<
+    import('./TokenManagerService').TokenManagerService
+  >({
+    id: 'core.tokenManager',
+  });
+
+  /**
+   * The service reference for the plugin scoped {@link UrlReaderService}.
+   *
+   * @public
+   */
+  export const urlReader = createServiceRef<
+    import('./UrlReaderService').UrlReaderService
+  >({
+    id: 'core.urlReader',
+  });
+}
